# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Waryaa Gaming Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /players:
    get:
      operationId: listPlayers
      tags: [players]
      summary: List all players
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of players
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Player"
    post:
      operationId: createPlayer
      tags: [players]
      summary: Register a player
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayerInput"
      responses:
        "201":
          description: Created player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
  /players/{id}:
    get:
      operationId: getPlayer
      tags: [players]
      summary: Get player profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Player profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
        "404":
          description: Not found
    patch:
      operationId: updatePlayer
      tags: [players]
      summary: Update player profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayerUpdate"
      responses:
        "200":
          description: Updated player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
  /players/{id}/match-history:
    get:
      operationId: getPlayerMatchHistory
      tags: [players]
      summary: Get player match history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Match history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Match"
...
  /news:
    get:
      operationId: listNews
      tags: [news]
      summary: List news articles
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [tournament, community, esports, federation]
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: News articles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NewsArticle"
    post:
      operationId: createNews
      tags: [news]
      summary: Create a news article
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsInput"
      responses:
        "201":
          description: Created article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsArticle"
  /news/{id}:
    get:
      operationId: getNews
      tags: [news]
      summary: Get a news article
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: News article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsArticle"
        "404":
          description: Not found
...
components:
  schemas:
...
    Player:
      type: object
      required: [id, username, rank, tournamentWins, winRate, matchesPlayed, matchesWon, points, isActive]
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        teamId:
          type: ["integer", "null"]
        teamName:
          type: ["string", "null"]
        rank:
          type: integer
        tournamentWins:
          type: integer
        winRate:
          type: number
        matchesPlayed:
          type: integer
        matchesWon:
          type: integer
        points:
          type: integer
        country:
          type: ["string", "null"]
        discordId:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        isActive:
          type: boolean
        badges:
          type: array
          items:
            type: string
        createdAt:
          type: string
    PlayerInput:
      type: object
      required: [username]
      properties:
        username:
          type: string
        displayName:
          type: string
        country:
          type: string
        discordId:
          type: string
        bio:
          type: string
    PlayerUpdate:
      type: object
      properties:
        displayName:
          type: string
        country:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
...
    Match:
      type: object
      required: [id, tournamentId, round, status]
      properties:
        id:
          type: integer
        tournamentId:
          type: integer
        round:
          type: integer
        roundName:
          type: ["string", "null"]
        status:
          type: string
          enum: [scheduled, live, completed, cancelled]
        participant1Id:
          type: ["integer", "null"]
        participant1Name:
          type: ["string", "null"]
        participant1Score:
          type: ["integer", "null"]
        participant2Id:
          type: ["integer", "null"]
        participant2Name:
          type: ["string", "null"]
        participant2Score:
          type: ["integer", "null"]
        winnerId:
          type: ["integer", "null"]
        winnerName:
          type: ["string", "null"]
        scheduledAt:
          type: ["string", "null"]
        streamUrl:
          type: ["string", "null"]
        tournamentName:
          type: ["string", "null"]
...
    NewsArticle:
      type: object
      required: [id, title, category, content, publishedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        category:
          type: string
          enum: [tournament, community, esports, federation]
        content:
          type: string
        excerpt:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        authorId:
          type: ["integer", "null"]
        authorName:
          type: ["string", "null"]
        publishedAt:
          type: string
        isFeatured:
          type: boolean
    NewsInput:
      type: object
      required: [title, category, content]
      properties:
        title:
          type: string
        category:
          type: string
          enum: [tournament, community, esports, federation]
        content:
          type: string
        excerpt:
          type: string
        imageUrl:
          type: string
        isFeatured:
          type: boolean
...